Move & add validation functions
